10test passed

+++ src/s21_sprintf/s21_sprintf.h
@@ -9,7 +9,7 @@ struct info {
   char spec;
 };
 
-char *itoa(int val, int base);
+char *itoa(long val, int base);
 void init_struct(struct info *mys);
 void cpf_width(const char *pF, struct info *mys, va_list input, int * i);
 void cpf_acc(const char *pF, struct info *mys, va_list input, int * i);
